Move routes to controllers/ dir.
