+++ lib/app.js
@@ -44,7 +44,7 @@ function runMaster() {
 
     var cpuCount = require('os').cpus().length;
     console.log('Serving directory "' + options.wwwroot + '" on port ' + options.port + '.');
-    require('./convert')().testGdal();
+    require('./controllers/convert')().testGdal();
 
     if (!exists(options.wwwroot)) {
         warn('"' + options.wwwroot + '" does not exist.');
@@ -155,7 +155,7 @@ var bypassUpstreamProxyHostsMap = (options.settings.bypassUpstreamProxyHosts || 
         return map;
     }, {});
 
-app.use('/proxy', require('./proxy')({
+app.use('/proxy', require('./controllers/proxy')({
     proxyableDomains: options.settings.allowProxyFor,
     proxyAllDomains: options.settings.proxyAllDomains,
     proxyAuth: options.proxyAuth,
@@ -164,9 +164,9 @@ app.use('/proxy', require('./proxy')({
     basicAuthentication: options.settings.basicAuthentication
 }));
 
-app.use('/proj4def', require('./proj4lookup'));            // Proj4def lookup service, to avoid downloading all definitions into the client.
-app.use('/convert', require('./convert')(options).router); // OGR2OGR wrapper to allow supporting file types like Shapefile.
-app.use('/proxyabledomains', require('./proxydomains')({   // Returns JSON list of domains we're willing to proxy for
+app.use('/proj4def', require('./controllers/proj4lookup'));            // Proj4def lookup service, to avoid downloading all definitions into the client.
+app.use('/convert', require('./controllers/convert')(options).router); // OGR2OGR wrapper to allow supporting file types like Shapefile.
+app.use('/proxyabledomains', require('./controllers/proxydomains')({   // Returns JSON list of domains we're willing to proxy for
     proxyableDomains: options.settings.allowProxyFor,
     proxyAllDomains: !!options.settings.proxyAllDomains,
 }));
@@ -182,7 +182,7 @@ if (serveWwwRoot) {
     initPaths.push(path.join(options.wwwroot, 'init'));
 }
 
-app.use('/init', require('./initfile')(initPaths, error404, path.dirname(options.configFile)));
+app.use('/init', require('./controllers/initfile')(initPaths, error404, path.dirname(options.configFile)));
 /*
 // For testing, simply reflects stuff back at the caller. Uncomment if needed.
 app.get('/reflect', function(req, res){
@@ -199,11 +199,11 @@ app.post('/reflect', bodyParser.urlencoded({extended: true, type: function() { r
 });
 */
 
-var feedbackService = require('./feedback')(options.settings.feedback);
+var feedbackService = require('./controllers/feedback')(options.settings.feedback);
 if (feedbackService) {
     app.use('/feedback', feedbackService);
 }
-var shareService = require('./share')(options.settings.shareUrlPrefixes, options.settings.newShareUrlPrefix);
+var shareService = require('./controllers/share')(options.settings.shareUrlPrefixes, options.settings.newShareUrlPrefix);
 if (shareService) {
     app.use('/share', shareService);
 }
