+++ lib/controllers/initfile.js
@@ -2,7 +2,7 @@
 'use strict';
 var express = require('express');
 var router = require('express').Router();
-var exists = require('./exists');
+var exists = require('../exists');
 var path = require('path');
 /**
  * Special handling for /init/foo.json requests: look in initPaths, not just wwwroot/init
