+++ lib/controllers/proj4lookup.js
