+++ lib/controllers/share.js
